multitask configs in get_config; augmented obs space with task_idx; tasks[0] copy in env constructor
