+++ configs/test/habitat_multitask_example.yaml
@@ -67,5 +67,4 @@ CHANGE_TASK_BEHAVIOUR:
 DATASET:
   TYPE: PointNav-v1
   SPLIT: val
-  # DATA_PATH: data/datasets/pointnav/gibson/v1/{split}/{split}.json.gz
   DATA_PATH: data/datasets/pointnav/habitat-test-scenes/v1/{split}/{split}.json.gz
